Idle handlers no longer defer to scene-graph parents.
Parents still wait for all children to be idle.

Before, if a map had multiple layers, for instance, `oneLayer.onIdle`
would wait for all layers and the map to be idle before triggering.
Now, `oneLayer.onIdle` will only wait for that layer and its children.
`mainMap.onIdle` will wait for all of the layers and their children as
before.

This is useful, for instance, if you have two tile layers and you wish
to trigger a function when one of the tile layers is fully loaded
without waiting for the second tile layer.  You can still wait for both
by using `map.onIdle`.